# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalogDTO": {
        "additionalProperties": false,
        "properties": {
          "as_of_date": {
            "description": "历史相似日期",
            "format": "date",
            "title": "As Of Date",
            "type": "string"
          },
          "distance": {
            "description": "距离分数，越小越相似",
            "minimum": 0.0,
            "title": "Distance",
            "type": "number"
          },
          "feature_set_version": {
            "description": "用于计算距离的特征版本（config/features/{version}.yaml）",
            "title": "Feature Set Version",
            "type": "string"
          },
          "features": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "description": "当时可见的 point-in-time 特征；某个特征当时算不出来就是 null，不补 0",
            "title": "Features",
            "type": "object"
          },
          "forward_return_1d": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "其后 1 个交易日的真实收益；历史数据不足以结算时为 null（不编造 0）",
            "title": "Forward Return 1D"
          },
          "forward_return_5d": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "其后 5 个交易日的真实累计收益；不足以结算时为 null",
            "title": "Forward Return 5D"
          },
          "horizon": {
            "$ref": "#/components/schemas/PredictionHorizon"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "horizon",
          "as_of_date",
          "distance",
          "features",
          "feature_set_version",
          "forward_return_1d",
          "forward_return_5d"
        ],
        "title": "AnalogDTO",
        "type": "object"
      },
      "AnalysisDTO": {
        "additionalProperties": false,
        "properties": {
          "analysis_type": {
            "$ref": "#/components/schemas/AnalysisType"
          },
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_cutoff": {
            "format": "date-time",
            "title": "Data Cutoff",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceDTO"
            },
            "title": "Evidence",
            "type": "array"
          },
          "horizon": {
            "$ref": "#/components/schemas/EventHorizon"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "model_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Provider"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "id",
          "symbol",
          "analysis_type",
          "direction",
          "horizon",
          "summary",
          "data_cutoff",
          "created_at"
        ],
        "title": "AnalysisDTO",
        "type": "object"
      },
      "AnalysisType": {
        "enum": [
          "document",
          "anomaly",
          "daily_brief"
        ],
        "title": "AnalysisType",
        "type": "string"
      },
...
      "BarsMetaDTO": {
        "additionalProperties": false,
        "properties": {
          "summaries": {
            "additionalProperties": {
              "$ref": "#/components/schemas/BarRangeSummaryDTO"
            },
            "title": "Summaries",
            "type": "object"
          },
          "timeframe": {
            "$ref": "#/components/schemas/Timeframe"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "timeframe",
          "total_count",
          "updated_at",
          "summaries"
        ],
        "title": "BarsMetaDTO",
        "type": "object"
      },
      "BarsResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/BarDTO"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/BarsMetaDTO"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "page",
          "meta",
          "request_id"
        ],
        "title": "BarsResponse",
        "type": "object"
      },
...
      "Direction": {
        "enum": [
          "positive",
          "negative",
          "neutral",
          "unknown"
        ],
        "title": "Direction",
        "type": "string"
      },
...
      "DocumentType": {
        "enum": [
          "announcement",
          "news"
        ],
        "title": "DocumentType",
        "type": "string"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "错误信封。HTTP 状态由 ``AppError.status_code`` 唯一决定（不在路由里硬编码）。",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ItemResponse_JobDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/JobDTO"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "ItemResponse[JobDTO]",
        "type": "object"
      },
...
      "ItemResponse_QuoteRefreshDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuoteRefreshDTO"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "title": "ItemResponse[QuoteRefreshDTO]",
        "type": "object"
      },
...
      "ListResponse_AnalogDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AnalogDTO"
            },
            "title": "Data",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "page",
          "request_id"
        ],
        "title": "ListResponse[AnalogDTO]",
        "type": "object"
      },
      "ListResponse_AnalysisDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AnalysisDTO"
            },
            "title": "Data",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "page",
          "request_id"
        ],
        "title": "ListResponse[AnalysisDTO]",
        "type": "object"
      },
      "ListResponse_DocumentDTO_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/DocumentDTO"
            },
            "title": "Data",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "data",
          "page",
          "request_id"
        ],
        "title": "ListResponse[DocumentDTO]",
        "type": "object"
      },
...
      "PageInfo": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "default": false,
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "PageInfo",
        "type": "object"
      },
...
      "PredictionHorizon": {
        "description": "预测目标（predictions.horizon / model_versions.target_horizon）。",
        "enum": [
          "today_close",
          "next_5d"
        ],
        "title": "PredictionHorizon",
        "type": "string"
      },
...
      "SnapshotResponse": {
        "additionalProperties": false,
        "description": "spec §7.2 的响应是**裸对象**（不带 data 包裹）；§7 又要求所有响应带 request_id，\n因此在裸对象上追加 request_id 字段（同时也在 ``X-Request-Id`` 响应头返回）。",
        "properties": {
          "is_current_universe_member": {
            "description": "false ⇒ 前端展示「已调出沪深300」（spec §3.1）",
            "title": "Is Current Universe Member",
            "type": "boolean"
          },
          "latest_anomaly_analysis_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Anomaly Analysis Id"
          },
          "latest_predictions": {
            "description": "各 horizon 的最新预测 id；无预测时为空数组",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Latest Predictions",
            "type": "array"
          },
          "market": {
            "$ref": "#/components/schemas/MarketDTO"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuoteDTO"
              },
              {
                "type": "null"
              }
            ],
            "description": "实时行情尚未取得时为 null；历史行情不填入此字段"
          },
          "relative_strength": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RelativeStrengthDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "name",
          "market",
          "is_current_universe_member",
          "request_id"
        ],
        "title": "SnapshotResponse",
        "type": "object"
      },
...
      "Timeframe": {
        "enum": [
          "5m",
          "1d"
        ],
        "title": "Timeframe",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "个人使用、仅沪深300成分股、OpenBB 统一数据网关、免费近实时数据、无自动交易。\n\n所有响应都带 request_id；列表使用游标分页（limit 默认 20、最大 100）。\n预测仅供研究，不构成投资建议。",
    "title": "A股 AI 研究助手 API",
    "version": "1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/stocks/{symbol}/analogs": {
      "get": {
        "operationId": "list_analogs_api_v1_stocks__symbol__analogs_get",
        "parameters": [
          {
            "description": "6 位 A 股代码，例如 600519",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "6 位 A 股代码，例如 600519",
              "examples": [
                "600519"
              ],
              "pattern": "^\\d{6}$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "horizon",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PredictionHorizon",
              "default": "next_5d"
            }
          },
          {
            "description": "默认 20，最大 100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 10,
              "description": "默认 20，最大 100",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_AnalogDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSUFFICIENT_DATA：历史样本不足，无法生成结果"
          }
        },
        "summary": "历史相似行情（有效候选 < 30 时关闭该功能 ⇒ 422）",
        "tags": [
          "stocks"
        ]
      }
    },
    "/api/v1/stocks/{symbol}/analyses": {
      "get": {
        "operationId": "list_analyses_api_v1_stocks__symbol__analyses_get",
        "parameters": [
          {
            "description": "6 位 A 股代码，例如 600519",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "6 位 A 股代码，例如 600519",
              "examples": [
                "600519"
              ],
              "pattern": "^\\d{6}$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "description": "document | anomaly | daily_brief",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AnalysisType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "document | anomaly | daily_brief",
              "title": "Type"
            }
          },
          {
            "description": "无填充 Base64URL 游标",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "无填充 Base64URL 游标",
              "title": "Cursor"
            }
          },
          {
            "description": "默认 20，最大 100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "默认 20，最大 100",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_AnalysisDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          }
        },
        "summary": "AI 解释（证据整条展开，绝不返回部分证据）",
        "tags": [
          "stocks"
        ]
      }
    },
    "/api/v1/stocks/{symbol}/analyses/refresh": {
      "post": {
        "operationId": "refresh_analyses_api_v1_stocks__symbol__analyses_refresh_post",
        "parameters": [
          {
            "description": "6 位 A 股代码，例如 600519",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "6 位 A 股代码，例如 600519",
              "examples": [
                "600519"
              ],
              "pattern": "^\\d{6}$",
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse_JobDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          }
        },
        "summary": "触发解释刷新（登记 analysis_refresh 作业，由 worker 执行）",
        "tags": [
          "stocks"
        ]
      }
    },
    "/api/v1/stocks/{symbol}/bars": {
      "get": {
        "operationId": "list_bars_api_v1_stocks__symbol__bars_get",
        "parameters": [
          {
            "description": "6 位 A 股代码，例如 600519",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "6 位 A 股代码，例如 600519",
              "examples": [
                "600519"
              ],
              "pattern": "^\\d{6}$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timeframe",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Timeframe",
              "default": "1d"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 240,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          }
        },
        "summary": "历史 K 线（与实时行情独立）",
        "tags": [
          "stocks"
        ]
      }
    },
    "/api/v1/stocks/{symbol}/documents": {
      "get": {
        "operationId": "list_documents_api_v1_stocks__symbol__documents_get",
        "parameters": [
          {
            "description": "6 位 A 股代码，例如 600519",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "6 位 A 股代码，例如 600519",
              "examples": [
                "600519"
              ],
              "pattern": "^\\d{6}$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "description": "announcement | news",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "announcement | news",
              "title": "Type"
            }
          },
          {
            "description": "无填充 Base64URL 游标",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "无填充 Base64URL 游标",
              "title": "Cursor"
            }
          },
          {
            "description": "默认 20，最大 100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "默认 20，最大 100",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_DocumentDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          }
        },
        "summary": "公告与新闻（按内容哈希去重）",
        "tags": [
          "stocks"
        ]
      }
    },
...
    "/api/v1/stocks/{symbol}/quote-refresh": {
      "post": {
        "operationId": "refresh_quote_api_v1_stocks__symbol__quote_refresh_post",
        "parameters": [
          {
            "description": "6 位 A 股代码，例如 600519",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "6 位 A 股代码，例如 600519",
              "examples": [
                "600519"
              ],
              "pattern": "^\\d{6}$",
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse_QuoteRefreshDTO_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          }
        },
        "summary": "为当前股票登记一次最新行情刷新任务",
        "tags": [
          "stocks"
        ]
      }
    },
    "/api/v1/stocks/{symbol}/snapshot": {
      "get": {
        "description": "行情过期时返回 stale；尚无实时行情时返回 200 + quote=null。",
        "operationId": "get_snapshot_api_v1_stocks__symbol__snapshot_get",
        "parameters": [
          {
            "description": "6 位 A 股代码，例如 600519",
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "description": "6 位 A 股代码，例如 600519",
              "examples": [
                "600519"
              ],
              "pattern": "^\\d{6}$",
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INVALID_ARGUMENT：代码、分页或时间参数无效"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INSTRUMENT_NOT_FOUND：证券不存在或不在 MVP 市场范围"
          }
        },
        "summary": "股票快照（行情、新鲜度、相对强弱、最新异动与预测）",
        "tags": [
          "stocks"
        ]
      }
    },
...
  }
}